rename component and directory from Type to PokemonTypes and refactor PokemonTypes component to dynamically render TypeBadge component from array of conditions
