+++ src/components/PokemonTypes/TypeBadge.jsx
